requirements.txt
Flask==2.0.2
Flask_Cors==3.0.10
mysql-connector-python==8.0.28
python-dotenv==0.19.2
requests==2.25.1
waitress==2.0.0
